修改数据库模型

+++ models/profile.go
@@ -3,8 +3,8 @@ package models
 import "time"
 
 type Profile struct {
-	ID          int8      `gorm:"id;primary_key"`
-	UserId      int8      `gorm:"user_id"`
+	ID          int64     `gorm:"id;primary_key"`
+	UserId      int64     `gorm:"user_id"`
 	Nickname    string    `gorm:"nickname"`
 	Email       string    `gorm:"email"`
 	Phone       string    `gorm:"phone"`
